requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
elastic-transport==8.17.0
elasticsearch==8.11.0
executing==2.2.0
fastapi==0.115.11
Flask==3.1.0
h11==0.14.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyspark==3.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
pyzmq==26.2.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.1
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
Werkzeug==3.1.3
